[FEAT] [subject: Community] 조회수 기능 추가, 이미지 수정 없이 글 작성 기능 추가[해결]

+++ src/main/java/com/example/canya/community/dto/CommunityResponseDto.java
@@ -13,7 +13,7 @@ public class CommunityResponseDto {
     private String communityTitle;
     private String communityContent;
     private String communityImage;
-    private Integer communityHitCount;
+    private Long communityHitCount;
     private String createdAt;
     private String memberNickname;
     private String memberProfileImage;
